Carrito Funcional

+++ perfil.php
@@ -114,6 +114,7 @@ if (session_status() == PHP_SESSION_ACTIVE && isset($_SESSION['usuario'])) {
                                     <th class="centrado">Cantidad</th>
                                     <th class="centrado">Precio</th>
                                     <th class="centrado">Total</th>
+                                    <th class="centrado">Stock</th>
                                     <th class="centrado">Acciones</th>
                                 </tr>
                             </thead>
@@ -201,81 +202,104 @@ if (session_status() == PHP_SESSION_ACTIVE && isset($_SESSION['usuario'])) {
                 });
             },
 
-            // Wait for the payment to be authorized by the customer
-
             // onAuthorize: function(data, actions) {
             //     return actions.payment.execute().then(async function() {
 
-            //         const respuesta = await fetch('PHP/MostrarCarrito.php');
-            //         const datos = await respuesta.json();
             //         let numProductos = 0,
             //             subtotal = 0,
             //             total = 0;
 
-
-            //         // Insertar en Ventas (Por mientras se iniciará para obtener el ID de venta ):                    
-            //         fetch('PHP/InsertarVenta.php', {
+            //         await fetch('PHP/InsertarVenta.php', {
             //             method: 'POST'
             //         });
 
             //         const respuesta2 = await fetch('PHP/ObtenerIDVenta.php');
             //         const id_venta = await respuesta2.json();
-            //         const ID = 0;
+            //         let ID = 0;
 
             //         id_venta.forEach(id => {
             //             ID = id.vta_id;
             //             console.log(ID);
-            //         })
+            //         });
 
 
-            //         datos.forEach(registro => {
+            //         const respuesta = await fetch('PHP/MostrarCarrito.php');
+            //         const datos = await respuesta.json();
 
-            //             if (registro.cardis_cantidad > registro.dis_existencia) {
-            //                 ejecutarProcedimiento(registro.dis_id, registro.cardis_cantidad);
+            //         console.log(datos);
 
-            //                 numProductos++;
+            //         const promesasDetalle = datos.map(async registro => {
+
+            //             if (registro.dis_existencia > registro.cardis_cantidad) {
+
+            //                 await ejecutarProcedimiento(registro.dis_id, registro.cardis_cantidad);
+
+            //                 numProductos += 1;
+            //                 console.log('Cuenta: ' + numProductos);
+
             //                 subtotal = (registro.cardis_cantidad * registro.dis_precioUnitario);
             //                 total += subtotal;
 
-            //                 // Insertar en Detalles:
+            //                 console.log('Cantidad: ' + registro.cardis_cantidad);
+            //                 console.log('Precio: ' + registro.dis_precioUnitario);
+            //                 console.log('SubTotal: ' + subtotal);
+            //                 console.log('ID Disco: ' + registro.dis_id);
+            //                 console.log('ID Venta: ' + ID);
+
             //                 const formDataDetalles = new FormData();
             //                 formDataDetalles.append('parametroCantidad', registro.cardis_cantidad);
             //                 formDataDetalles.append('parametroPrecio', registro.dis_precioUnitario);
             //                 formDataDetalles.append('parametroSubTotal', subtotal);
             //                 formDataDetalles.append('parametroIDDisco', registro.dis_id);
             //                 formDataDetalles.append('parametroIDVenta', ID);
-            //                 fetch('PHP/InsertarDetalle.php', {
+
+            //                 return fetch('PHP/InsertarDetalle.php', {
             //                     method: 'POST',
             //                     body: formDataDetalles
             //                 });
-
-
-            //                 eliminarDelCarrito(registro.cardis_id);
             //             }
-            //         })
+            //         });
 
-            //         // Modificar en Ventas el registro (Total y num de productos):                    
+            //         await Promise.all(promesasDetalle);
+
+            //         console.log('Total: $' + total);
+            //         console.log('Productos: ' + numProductos);
+            //         console.log('ID: ' + ID);
+
             //         const formDataVenta = new FormData();
             //         formDataVenta.append('parametroTotal', total);
             //         formDataVenta.append('parametroNumProductos', numProductos);
             //         formDataVenta.append('parametroID', ID);
-            //         fetch('PHP/ActualizarVenta.php', {
+
+            //         await fetch('PHP/ActualizarVenta.php', {
             //             method: 'POST',
             //             body: formDataVenta
             //         });
 
+            //         const respuesta3 = await fetch('PHP/MostrarCarrito.php');
+            //         const datos3 = await respuesta3.json();
+
+            //         const promesasDetalle2 = datos3.map(async registro => {
+            //             if (registro.dis_existencia > registro.cardis_cantidad) {
+            //                 await eliminarDelCarrito(registro.cardis_id);
+            //             }
+            //         });
+
+            //         await Promise.all(promesasDetalle2);
             //         window.alert('¡Pago realizado!');
             //         window.location = "perfil.php";
+
+
+
 
             //     });
             // }
 
             onAuthorize: function(data, actions) {
                 return actions.payment.execute().then(async function() {
-
-                    let numProductos = 0,
-                        subtotal = 0,
-                        total = 0;
+                    let numProductos = 0;
+                    let subtotal = 0;
+                    let total = 0;
 
                     await fetch('PHP/InsertarVenta.php', {
                         method: 'POST'
@@ -290,41 +314,41 @@ if (session_status() == PHP_SESSION_ACTIVE && isset($_SESSION['usuario'])) {
                         console.log(ID);
                     });
 
-
                     const respuesta = await fetch('PHP/MostrarCarrito.php');
                     const datos = await respuesta.json();
 
                     console.log(datos);
 
-                    const promesasDetalle = datos.map(async registro => {
+                    const promesasDetalle = datos.map(registro => {
+                        if (parseInt(registro.dis_existencia) > parseInt(registro.cardis_cantidad)) {
+                            return ejecutarProcedimiento(registro.dis_id, registro.cardis_cantidad)
+                                .then(() => {
+                                    numProductos += 1;
+                                    console.log('Cuenta: ' + numProductos);
 
-                        if (registro.dis_existencia > registro.cardis_cantidad) {
+                                    subtotal = (registro.cardis_cantidad * registro.dis_precioUnitario);
+                                    total += subtotal;
 
-                            await ejecutarProcedimiento(registro.dis_id, registro.cardis_cantidad);
+                                    console.log('Cantidad: ' + registro.cardis_cantidad);
+                                    console.log('Precio: ' + registro.dis_precioUnitario);
+                                    console.log('SubTotal: ' + subtotal);
+                                    console.log('ID Disco: ' + registro.dis_id);
+                                    console.log('ID Venta: ' + ID);
 
-                            numProductos += 1;
-                            console.log('Cuenta: ' + numProductos);
+                                    const formDataDetalles = new FormData();
+                                    formDataDetalles.append('parametroCantidad', registro.cardis_cantidad);
+                                    formDataDetalles.append('parametroPrecio', registro.dis_precioUnitario);
+                                    formDataDetalles.append('parametroSubTotal', subtotal);
+                                    formDataDetalles.append('parametroIDDisco', registro.dis_id);
+                                    formDataDetalles.append('parametroIDVenta', ID);
 
-                            subtotal = (registro.cardis_cantidad * registro.dis_precioUnitario);
-                            total += subtotal;
-
-                            console.log('Cantidad: ' + registro.cardis_cantidad);
-                            console.log('Precio: ' + registro.dis_precioUnitario);
-                            console.log('SubTotal: ' + subtotal);
-                            console.log('ID Disco: ' + registro.dis_id);
-                            console.log('ID Venta: ' + ID);
-
-                            const formDataDetalles = new FormData();
-                            formDataDetalles.append('parametroCantidad', registro.cardis_cantidad);
-                            formDataDetalles.append('parametroPrecio', registro.dis_precioUnitario);
-                            formDataDetalles.append('parametroSubTotal', subtotal);
-                            formDataDetalles.append('parametroIDDisco', registro.dis_id);
-                            formDataDetalles.append('parametroIDVenta', ID);
-
-                            return fetch('PHP/InsertarDetalle.php', {
-                                method: 'POST',
-                                body: formDataDetalles
-                            });
+                                    return fetch('PHP/InsertarDetalle.php', {
+                                        method: 'POST',
+                                        body: formDataDetalles
+                                    });
+                                });
+                        } else {
+                            return Promise.resolve(); // Promesa vacía si no se cumple la condición
                         }
                     });
 
@@ -344,14 +368,24 @@ if (session_status() == PHP_SESSION_ACTIVE && isset($_SESSION['usuario'])) {
                         body: formDataVenta
                     });
 
-                    datos.forEach(registro => {
-                        eliminarDelCarrito(registro.cardis_id);
+                    const respuesta3 = await fetch('PHP/MostrarCarrito.php');
+                    const datos3 = await respuesta3.json();
+
+                    const promesasDetalle2 = datos3.map(registro => {
+                        if (parseInt(registro.dis_existencia) > parseInt(registro.cardis_cantidad)) {
+                            return eliminarDelCarrito(registro.cardis_id);
+                        } else {
+                            return Promise.resolve(); // Promesa vacía si no se cumple la condición
+                        }
                     });
 
+                    await Promise.all(promesasDetalle2);
+
                     window.alert('¡Pago realizado!');
-                    // window.location = "perfil.php";
+                    window.location = "perfil.php";
                 });
             }
+
 
 
         }, '#paypal-button-container');
